Split AdminEquipmentTypes.jsx
