+++ indico/modules/rb_new/client/js/modules/admin/EquipmentTypesPage.module.scss
